requirements.txt
langchain
langchain-community
sentence-transformers
pypdf
flask
langchain-pinecone
python-dotenv
langchain-huggingface
pinecone